64-bit fixes for sipXmediaLib.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@10922 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/include/mp/MpAudioOutputConnection.h
@@ -318,7 +318,7 @@ protected:
 
      /// Call this when driver become ready for the next frame.
    static
-   void tickerCallback(const int userData, const int eventData);
+   void tickerCallback(const intptr_t userData, const intptr_t eventData);
      /**<
      *  @param userData - (in) Contain pointer connection it is associated with.
      *  @param eventData - (in) contain 0 for now.
